requirements.txt
asgiref==3.8.1
Django==5.1.1
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.9
sqlparse==0.5.1
whitenoise==6.7.0
djangorestframework==3.15.2
selenium==4.27.1